Embrace "Isotope" and "BAD-WORD!" meanings
The terminology of saying something is an "isotope" (as if another
version would not be) isn't quite correct.  A family of isotopes are
all isotopes of each other.

However, as null isotopes have merged into be the BAD-WORD! ~null~
isotope, it has come to make sense that speaking of the version that
cannot be put in blocks as being the (less common) "isotope".  So
the plain version could be called a BAD-WORD!.

This replaces the `Init_Curse_Word()` with `Init_Isotope()`, as the
term "curse word" is not going to be used in userspace.

It also gets rid of the term "light nulled" and instead just speaks
of the "null isotope", which is known to be a BAD-WORD! with label
"null" and the isotope flag set.