Handle `join 'a/b/c 'd/e/f` case
The usermode implementation of JOIN has a lot of work to be done, and
it may not be desirable to allow `join 'a/b/c 'd/e/f`...because slashes
might need to be lined up as `join 'a/b/c '/d/e/f`.

For this reason the implementation had not been updated.  However, at
the moment fixing the tests is a priority...so the best way to
document the question is just to add the code to JOIN to handle it.