Use "/ONLY" semantics for group insertion in test
The reason at the moment why GROUP! is not allowed to be appended to
a series without being quoted is that it is evaluative.  This gives an
answer for `append [a b c] '(d e f)`... it is neither `[a b c d e f]`
nor `[a b c (d e f)]`.