Make virtual binding to new context non-expandable
There are historical bugs pertaining to getting the context of a
loop variable in something like a FOR-EACH, and then expanding the
context out from under pointers held by the FOR-EACH implementation.

The iterator can either not lock those variables, or the context can
be locked against expansion.  There had been code that locked against
expansion previously but it seems it was deleted, but a test was
catching it.

This fixes that as part of fixing up the test breakages.