Remove clearing of stale flag from GROUP! evaluations
A line commenting that it was to enable `any [(10 elide "hi")]` was
clearing a stale flag, but it's not clear that clearing that flag
worked...while it definitely broke the feature of things like:

    >> case [(elide "hi") true [1]]
    == 1

The presence of the GROUP! is not supposed to interfere, and that
should act as you had written `case [elide "hi" true [1]]`.  CASE
preloads the output with NULL and uses the stale flag to decide if
the cell was overwritten.

It doesn't seem to break any tests to remove that flag clearing, and
it makes the CASE scenario work.  So this removes the line.