Correct handling of isotopes in HELP stress test
The HELP stress test had been failing, and along with several other
lapses in testing it was allowed to go on while some things were being
pinned down.

Now that the reasons HELP and DUMP themselves were failing in some
cases have been fixed, an error in the stress test's attempt to avoid
BAD-WORD!s was noticed.  It is trying to "meta-get" a word holding the
name of a variable, instead of GET/ANY on that variable.

This fixes that, allowing the tests to work again.