Clarify why IS_TRUTHY() tolerates null isotopes
Technically speaking, users need not think of whether the ~null~
isotope is "truthy or falsey"...because most functions they would
call (like DID) won't see the isotope since they take normal
parameters...

But internal to the implementation, there is code like:

     >> every x [1 2 3] [if x = 2 [null]]
     == [1 3]

The body result for x = 2 is a ~null~ isotope.  To say whether that
isotope is "fundamentally truthy" or that functions like this
"make a special exception for null isotopes" is splitting hairs.

It seems easier to just tolerate them here vs. asserting it never
sees isotopes, and having a separate version of IS_TRUTHY().