Change @ ~null~ to produce true NULL
This changes the @ operation...which is intended as a convenience for
API users...to make the unusual tradeoff of not accepting any BAD-WORD!
values besides ~NULL~, and turning it into true NULL.  All other values
are passed as-is.

As a result, it's less likely to be the case to see problems from
things like:

    REBVAL *var_might_be_null = nullptr;

    rebElide("@", var_might_be_null, "else [print {This should run}]");

The API splices in a ~null~ to the execution stream in this case as
a convenience.

For full generality, users should use rebQ(var_might_be_null), but in
many situations this is more wordy as well as more slow.

An alternative that would preserve the value ~null~ could be to use
the THE operation.  However, this would render a "true null" as the
BAD-WORD! null in the case above.