Allow comparison of HANDLE!
The HANDLE! type in R3-Alpha had a bug where all handle types compared
as equal.  Rather than write a comparison function for them, Ren-C
punted on it...since it's not clear exactly what a generic notion
of EQUAL? would mean for a handle...especially with the difference
between shared handles and non-shared ones.

But the error was triggered when doing memberwise comparison on objects
representing ports, during a FIND function.  This actually raises
questions of whether PORT!s should have the FIND done by something
doing memberwise comparison or a behavior more like SAME?.  This just
points out to how informal the answer to questions of "sameness" and
"equality" were in R3-Alpha.

Rather than figure that out right now, this kicks the can down the road
a little farther by coming up with a way to compare handles so that
FIND works when looking for PORT!s in a block.