Fix off-by-one error in HELP descriptions
If a function did not have a RETURN: specification explicitly, then
there was an issue where help strings would be off by one.

In addition to resolving that issue, this adds a new "first" of having
some minimal testing that the help strings in the auxiliary object
(from which HELP is generated) are correct.  It adds a test for the
AUGMENT function as well, that it keeps the existing parameter
descriptions and adds its own correctly.