Update SKIP's return result to permit NULL
Because SKIP does not sound like it should use or synthesize any input
values, it was changed to be invisible.  The `<any>` instruction is used
to match a single atom of input and return it.

However SKIP can fail, and if it does it returns NULL.  This was not
in the type spec.

This corrects that issue and adds tests which include those for CHANGE
when used with OPT.