Date comparison errors if TZ or time presence mismatch
A DATE! which does not have a time zone or time cannot be compared
with one that has the more precise information.  It is misleading to
pretend that an unspecified time zone acts as GMT or to be assumed as
local, so users should explicitly standardize the information before
doing compares.

This brings the feature in line with tests written by Brett several
years ago which have been on a lingering branch.