Eliminate META-XXX! Branch Forms
The SYM-XXX! forms had been introduced to be able to subvert null
isotopification.  That transferred over to the META-XXX! forms, but
it's an inappropriate behavior...they should be "meta branches".

This is a transitional commit which backs out the behavior in order
to find the (almost certainly few) uses of META-XXX! types to give
"as-is" branching.  Once that is done, the concept of what ^branch
forms do can be revisited.