Distinguish UNQUOTE and UNMETA
The UNQUOTE operation had been twisted up a bit in order to let it
serve as the way for undoing meta operations.  But with the word META
coming into play as the English form for what ^ does, it becomes
somewhat natural to put that functionality under UNMETA and then keep
UNQUOTE as a more pure operation that just removes quote levels.