Allow SPECIALIZE with ~unset~ isotopes
This introduces a unique tag that SPECIALIZE uses to check by identity
in order to see if a slot is unspecialized.  It is known internally
only to SPECIALIZE.

Consequently, it's possible to specialize meta parameters with unset
isotopes.

To demonstrate this working, the UNSET function is changed to a
specialization of the SET function.