Make CALL of BLOCK! do a COMPOSE, allow WORD!s etc.
CALL can take a string or a block on all platforms.  If the execution
API is block based it will convert strings, and if it is string based
it will convert blocks.

Often when a BLOCK! is used some form of REDUCE or COMPOSE is needed.
If we biased the CALL interface for BLOCK! to use COMPOSE, then it is
possible to let things that would appear on shell command lines not
in quotes (like `--version` or `/Wd1020` or `foo/file.txt`) to be
converted to text.  This provides a more convenient interface for
execution that looks more like a command line would look...escaping
calculations into GROUP!s.