Fix inbuf_pos bug in CALL, wait on process
There was a variable that was being reinitialized to 0 each time through
a loop (inbuf_pos) when it needed to persist across loop iterations.
The bug had not shown itself since the synchronous input write generally
succeeded in one call historically.  But if it didn't it would wind up
writing the input from the beginning each time.

Also adds the process handle to the wait list with the file handles.
This seems a cleaner idea than needing to try writing to file handles
to find out that they have been closed.