READ-LINE hacks to try and do *some* piped stdin
R3-Alpha was not designed to handle stdio piping, and the port model
overall had a very unarticulated model.

This adds some nasty hacks to try and coax the system to interact with
the piped I/O of CALL to script the input to ASK reading lines at
a time.