Make COMBINATOR a native with its own dispatcher
The COMBINATOR generator exists for two reasons...one to standardize
the function interface for parser combinators, and secondly to give
a hook into each time one parser calls another.  This allows for
tracing as well as things like measuring how far the parse has gotten.

The usermode implementation of combinator was a complex enclosure of
a function, that this speeds up with native code.