Eliminate NATIVE/BODY Idea
There was an early Ren-C idea to augment natives in their specs with
a possible alternative expression of the body in usermode code.  Nice
in theory but has several flaws and reasons it's not as good as it
sounds, so it never got much use...and just made the processing code
confusing.

This removes the idea.  However, the "body" slot of natives was used
by various forms of natives (JavaScript natives for their JS source,
TCC natives for their C source, Generic functions for holding their
"VERB", etc.)  Those functions still exist so the slot still exists...
just not the feature for NATIVE.