Make empty moldings not count in delimiting
There are historical issues in Rebol where things like empty blocks
will (usually) throw off the printing process:

    rebol2>> print ["some" [] "stuff"]
    some  stuff

This makes things like BLOCK! less useful as a way of gathering pieces
of input, since the empty case is still treated as if something is
actually there.

Because it's not possible to predict in advance whether a molding
operation will produce output, this keeps track of the position before
a delimiter is added to the mold stream.  Then it marks the position
after the delimiter is molded.  If this is the position the mold buffer
is still at after the item is formed, then it's considered a case
for rollback.