Change GET-BLOCK! back to reducing behavior
The behavior of GET-BLOCK! was questioned as to whether it should act
as a GET of a BLOCK! would, in the classical sense.

For a time it was tried to see if a GET-BLOCK! would give more value
as a non-evaluative type...where the recipient could make the decision
on what it meant.  That idea was superseded by the ^META block which
could annotate a quote mark onto a block received by the callee...
which could serve as a signal if need be.  (Also, the @[...] type
will be introduced and will be another inert block type.)

So making GET-BLOCK! reduce has more value overall.