Change vaporization to empty block, not empty string
It seems to reason that DELIMIT must always put delimiters when an
empty string is involved.  Cases like comma-delimited formats do not
want empty strings to not be delimited.

This changes the vaporization to be a detail of the empty block case,
and not done by rolling back notice of empty FORM operations.