Ease UPARSE prohibition rule for WORD! fetches
Only prohibit WORD! or INTEGER! for now; basically anything that can
potentially take arguments to the combinator (word could be a rule,
integer takes the rule to repeat).

Error messages guide you to use a BLOCK! instead of WORD! or to use
REPEAT instead of the integer.

Fixes a mistake in the delimit tests.