Add wonky READ-CHAR function for whitespace tests
The sketchy R3-Alpha I/O subsystem (which is thrown out completely in
the Wasm build) was not written to support piping of standard input
or output very well.

However, for doing automated tests of the whitespace interpreter, there
is a need for character-by-character input.  This tries to hack
something together that works for that, but only in piped I/O for now.