Separate REB_TS_XXX bits into PARAM_FLAG_XXX bits
Historically, function parameters were represented by a TYPESET!.  This
had 64 bits of type flags, and then had to hold a symbol for the name
of the parameter.  That exhausted the space in the cell.

Ren-C needed more flags on parameters to represent things like
constness, endability, noop-on-blank, etc.  This meant those flags
were chewed out of the 64 bits for types, reducing the total number
of available datatypes (under these rules).

However, a change to move the keys of contexts into a more minimal
form allowed the symbol component of a parameter to be moved out into
the "exemplar" of actions.  This freed up more bits in typesets,
which could be allocated specifically to parameter flags.

With additional changes of other "pseudotype" usage, this should
recover bits from the 64 to be used for base types!