First Draft of Better "Rollback" Concept for COLLECT
This rethinks the way COLLECT works.  At first it had been a BLOCK! that
was kept in the PARSE state, which was rolled back whenever any alternate
in a BLOCK! rule failed.  This was then changed to roll back whenever
any rule failed...but having it hooked into the BLOCK! alternates had
handled the fact that BLOCK! didn't want *all* successful parsers it
called to contribute, only the ones that were part of a completely
successful alternate group.

That raised the question of how exactly the category of parsers like
BLOCK! could have a hand in management of which accumulated values would
be applicable to the result, while other parsers like SOME or WHILE or
BETWEEN could just go with the idea of rolling back on failure.  What
this commit does to achieve that goal is to make it so that if a
combinator mentions a `pending:` result value, it is assumed that it
will handle all the propagation and merging of synthesized collected
results.  If it does *not* mention `pending:` in its spec, then the
parsers it is given are wired up to themselves not have pending results
but instead manage those through the straightforward "success means
keep, failure means drop, and successful results are compounded in the
order the parsers are called".

It's a proof of concept, and by no means done efficiently...this is just
the first formulation that passes all the tests, and accomplishes the
stated goal.  It also is not generalized yet; the mechanism is being
applied just to COLLECT.  But the concept is that the gathered material
would contain tagged elements of some kind (possibly making exceptions
for common cases... QUOTED! items for COLLECT material, GROUP!s for
delayed executions, BLOCK! for EMIT SET-WORD! + value, etc.)