Re-introduce the @ datatypes
The use of @ for "modal parameters" eventually evolved into a mechanism
that was not related to the parameter, but to @ itself...which became
a "meta" operation in the evaluator.  This was deemed to make more
sense taking the ^ character.

However, there was not room in the type list to fit the @ variants with
no evaluative behavior.  That room has been made now, so the types
have been brought back.

   @word
   @pa/th
   @tu.p.le
   @[bl o ck]
   @(gr o up)