Rename REB_P_XX to more obvious PARAM_CLASS_XXX
Due to the scarcity of bits in typesets (64 bits + symbol) the only
place to put the parameter class for some time was in the KIND3Q
byte slot.  Hence the class was kept consistent with the REB_XXX
naming convention.

However, when the symbol moved out into a keylist then 32 bits worth
of space opened up...which could provide space for the parameter class
as well as a lot of flags.  With this change, the clearer naming of
PARAM_CLASS_XXX makes much more sense.