Change AS FRAME! aliasing of action to EXEMPLAR OF
The idea of being able to alias actions as frames and vice versa is
under consideration...where such an operation would take ownership of
the frame as read-only and expire extant references.

Either way, the AS would not likely be based on giving back the
exemplar.  This steps back from that presumption by making a separate
reflector for the purpose, to allow more room for AS mechanics to be
defined differently.