Eliminate PARAM_CLASS_LOCAL
With the unification of specialization frames and "exemplars", there
is no longer a "local" parameter class.  Locals act like any other
specialized parameter, they are just specialized to ~unset~.

This means that the existence of PARAM_CLASS_LOCAL was largely just
confusing; it was used in function spec analysis only.  So this
switches that to use a boolean instead in the analysis state and drops
the suggestion that there is such an actual parameter class.