Sort out frame phase visibility issues
The concept behind frame "phases" is that the same memory can be used
for multiple parts of the composition of a function.  So an ADAPT of
a SPECIALIZE of an AUGMENT of a FUNC can use the same list of variables
for the entire call.

However, each layer has to consider the inner layer's non-public
interface (locals, specialized values) to be invisible.  This way the
implementation details (such as having a local variable X or a
specialized out variable X) would not stop something like an AUGMENT
from adding an /X refinement.

This concept had been percolating around in the design for many years,
and tests had been written for what the theoretical behavior should be.
However, it was never successfully implemented...until now.

The basic approach just heeds the phase as originally intended.  The
twist is that the hiding of parameters is now done based on noticing
when a cell is unspecialized (e.g. a TYPESET! without a PARAM_CLASS)
and saying that the only time specialized cells (including locals)
are visible is when the phase matches the exemplar.  In order to
handle the case of copied paramlists that would give too much visibility
under this rule, the copied exemplar marks the cells that should not
be visible with VAR_MARKED_HIDDEN.

To make it easier to follow these rules, an enumeration struct and
routine are added.  It does not enumerate in partials order (e.g. the
order of the paramlist) but in the order of the actual variables in
the FRAME!.  This may or may not be a good idea.

Special handling is given to RETURN for visibility, but that needs a
larger amount of review.