Keep extant FRAME! pointers valid across ENCLOSE
Previously "to be safe" all extant frame pointers were invalidated at
the moment of enclose, since the variables were being re-thought as
an executable frame.

This did not allow "frame tunneling", e.g. a higher level version of
a function could not pass access to the frame to lower levels.  That
may or may not be a good idea.  But this change lets it be tried out.

The old code that invalidated the extant frames is kept in case this
turns out to be a bad idea.