Eliminate REB_META_PATH branch form
It may be that meta-branches are interesting, e.g.

    >> if foo ^[1 + 1]
    == '2

    >> if foo ^[~unset~]
    == ~unset~
         ; ^-- not an isotope

But this has not yet been solidified as desirable, and it's not what
REB_META_PATH was doing (it was a holdover from when META was SYM and
was behaving as @pa/th...simply an oversight that it wasn't deleted.)
Revisit when it seems like it might be a priority.