Add fake /PART option to UPARSE
The Red tests include tests of the PARSE/PART function.

In order to run those tests, this makes UPARSE internally use COPY/PART
if the option is used.  That is not what the intent of /PART is, but
it's good enough to pass those tests.

General questions of slicing series are being investigated, which may
be a better idea than troubling the combinator authors with honoring
a /PART parameter.