Deprecate plain END and HERE in UPARSE
UPARSE transitioned to using TAG! combinators for nouns, which offers
the benefit that names like END or HERE can be kept for variables.
This has turned out to be very appealing--so much so that it makes
sense to be the default unless you explicitly override it with your
own END and HERE combinators.

The method can extend to things like `<input>` for getting the input
series, or `<file>` and `<line>` for the current file and line being
parsed (as those features evolve).