Break %parse-tests-from-red into the per-combinator files
This breaks down the monolithic tests from %parse-test.red so that
they are in files for the combinators used.

It's an inexact science but it is generally more useful than organizing
tests based on what input type is being processed...because the pattern
across tests for the same combinator is more clear.