Bring back @ as "circling" facility for multi-return
For a time, the ability to mark parameters as the intended return value
was being done with ^META words, even though it was supposed to be done
with @THE words.  But the number of datatypes was restricted so @THE
words did not exist for a while.

Since they are back, this puts back the idea that @WORD is used in a
multi-return to say which multi-return should be the overall result
of the expression.