Add BREAK and STOP features for looping combinators
This is a new spin on the BREAK/ACCEPT/REJECT combinators that instead
says that BREAK should not match, and STOP does match.

They only work with loops, and as such make more sense than the
historical constructs.

The implementation is not ideal; it uses a manually built "loop stack"
tracked in the UPARSE state.  Clearer design systemically surrounding
the THROW mechanism and how local state in combinators might be
gracefully unwound is needed.

Includes both the usermode and native implementation for SOME, to prove
a native combinator can also participate.