Make different quote levels of values non-lax-equal
Historically '''a and a and 'a would all be considered equal under
non-strict rules.  You had to use == to tell them apart.

This was based on trying to take the behavior of historical Rebol's
`'a = a` and `not 'a == a` and extend it across generic quoting.
However, the idea that these values would be equal is not something
that has good properties, e.g. in PARSE one does not want to see
a '| as being equal to |.  One is a separator, and the other literally
matches | in the input.

This transitions to saying that differing quote levels do not match
even in lax equality.