Fix Parse_One_Rule() bug of going out of range
The Parse_One_Rule() function is supposed to be operating in terms of
the `pos` parameter passed in, not the parse position...this is because
when it is used in a TO or THRU rule, the parse position is not
advanced on each test step.

But the parse position was what was being checked for being at the
end of the series, not the advancing `pos`.  This means a TO or THRU
rule which hit the end could go out of range without checking.

The only reason this wasn't discovered earlier is probably that since
R3-Alpha did not allow complex rules as arguments to TO or THRU, there
wasn't much code that was exercising this.