Add support for LIT-WORD! and LIT-PATH! hacks in uparse
The LIT-WORD! and LIT-PATH! are "type constraints"; they're actually
old names for QUOTED! that are being faked up as constrained forms
of quoted word and quoted path.

There are some ideas floating around about how generic type constraints
might be implemented.  But until those are solidified, these legacy
entries are implemented because they are used by the QUERY dialect.

This is a temporary state of affairs.