Allow use of COMBINATOR that aren't in /COMBINATOR map
UPARSE can be parameterized with a map from keywords to combinators
that will apply regardless of what is in scope.

This makes it so if you use a WORD! that is not in that keyword map
but it looks up to a combinator, it will run that combinator as if
it were a keyword.  This allows the creation of local combinators
(similar to Brett Handley's PARSING-AT).