Allow NULL to be used in endable args run by frame
If you mark an evaluative parameter as `<end>` but don't also mark
it as `<opt>`, that means a NULL signal will be guaranteed in calling
to have come from actually being at the end of code...vs. being passed
as NULL (or some function returning NULL).

Currently that distinction can't be made when invoking a function via
a FRAME!.  NULL is NULL and the "endish" null mechanic used as a hack
to accomplish the feature in evaluator function application isn't
available.

This suggests that some other mechanism is going to be needed to
achieve the `<end>` mechanic, or it simply will not be able to be
distinguished from NULL when you are using a non-meta parameter.
That decision is deferred for the moment--this is just to keep code
working for now.