Make COLLECT/KEEP use QUOTED! items in pending array
The pending array is aiming to be a kind of amalgamation of various
value types for constructs to sort and filter through.

As a first cut at developing the pending feature, the array was solely
used for COLLECT.  It is presumed to be a very important client of the
feature, but not the only one.

This makes it so that COLLECT just puts QUOTED! items into the pending
array, and only assumes those items are for it.  That leaves every
other value type (BLOCK!, GROUP!, TEXT!, OBJECT!, etc.) available.

Ultimately it may be a tagged type like EVENT! would be a good fit for
some things, but OBJECT! with a `combinator:` key could cover an
ecology of items in the near term...albeit somewhat inefficiently.