Rewrite GATHER/EMIT to use the pending mechanism
This makes EMIT put BLOCK!s containing SET-WORD! and value pairs
into the pending stream, and then filter those out in order to
implement a GATHER of an object.

That allowed the old mechanic of trying to mark a "gathering state"
to be rolled back to be deleted...and makes GATHER inherit the
appearance of rollback in functions that have custom pending
aggregation.

The MAXMATCH demo is modified to show both COLLECT and GATHER mixing
their values together.  The QUOTED! values in the pending stream are
pulled out by COLLECT, while the BLOCK! values are taken by GATHER.