Fix out of bounds array access in SET
An address sanitizer test had not been run with R3_ALWAYS_MALLOC since
the change for arrays to not have terminal END nodes.  Some code in
SET which overstepped the boundary due to testing for that non-present
END triggered a sanitizer trap.

This fixes it, although SET is likely to be pared down to not run on
BLOCK!s...due to the introduction of UNPACK which solves many problems
with that idea.