Make return void and return none work in natives
The `return: <void>` and `return: <none>` designators (and their
historical predecessors) indicated that the standard function dispatcher
not be used, but one that would suppress the return value and either
make the result vanish or be written over opaquely.

Machinery for handling this could not apply to natives, because the
natives are their own dispatchers.  So the idea was that natives would
just use ordinary type blocks.

However, `return: []` does not currently distinguish between ~none~
isotopes and other forms of BAD-WORD! isotopes.  They are all
considered to be legal.

This commit switches it around so that these designators can still
be used on natives, though the behavior is still just ordinary type
checking.  At the specification level, it makes it clearer.