Get rid of "auto gather" in UPARSE
There was an idea that turned out to be bad, which was that if EMIT
was used inside of UPARSE with no associated GATHER the result was
assumed to be intended to be gathered.

This was disproven by a simple example from giuliolunati:

    uparse "ab" [collect [emit x: "a" keep "b"]] => ?

There doesn't seem to be any clear boundary to when stray emits should
consider it "proper" to override the seeming intended synthesized
result of a block.  So the feature is scrapped.

Someone doing a number of repetitive tasks where they have to say
GATHER more than they want could create their own UPARSE wrapper that
does what they want.