Add and use RETURN in spec for all functions
This change is a prelude to a rule which would say that if you have
RETURN: in your function spec then you must use it.  This would make
it simpler because RETURN would then be the only means to type check,
as well as to guarantee that a hooked RETURN function would be used
on all call paths.