Restore RETURN to being public slot on interface
Because RETURN was being filled by a function it was thought of as
being one of the function's "locals", in order to prevent it from
getting preloaded in a FRAME!.  However, fundamentally speaking the
properties associated with the RETURN have to be public in order
for the HELP and such to see it.

Also, some concepts are being thought about where RETURN: could be
filled with a WORD! on the public side of the interface, and that
word would be assigned in the process of the execution...just like
other outputs.