Change DECODE-URL to use UPARSE and return OBJECT!
While converting DECODE-URL to use UPARSE, the seeming ideal mechanism
to use was GATHER/EMIT and make an OBJECT!.  Since Red's DECODE-URL
returns an OBJECT! this was thought to be probably a fine idea.

But the historical return result of DECODE-URL was not an OBJECT, but
a BLOCK!.  This was done in order to allow the block to be used as
the argument of overrides in a construction spec.

Changing to an object shook up many bits of some very hacky code
scattered around.  This tries to keep that code working, though it's
up for a shakeup anyway.

One aspect of the change here is that the DECODE-URL returned object
has a fixed set of fields, and sets them to NULL if they are not
in use.  It also uses BAD-WORD! isotopes when something is a
contingent field...e.g. it shouldn't be accessed unless you've
noticed that a field it needs is present or not (like not reading
the password until you know there's a user field).