Preserve BLOCK! scheme specs in spec.ref
When a PORT! is opened via a BLOCK!, the concept is apparently that
the block can be dialected.  The httpd server takes advantage of that,
and expects to find the original dialect block in spec.ref -- even
though that block has to be valid for MAKE OBJECT!.

This conforms to that expectation for now.