Add new-style ANY combinator
This reclaims the ANY keyword in UPARSE for a combinator that will
treat a block it gets as a set of alternatives, without the need to
split them with | like in a normal BLOCK! rule.

    >> uparse? "ab" [any (["a" "b"])]
    == #[true]

It has to take its BLOCK! as a synthesized product, because if it were
just passed normally there'd be no way to override the semantics of
the typical BLOCK! combinator.  Hence the GROUP! around it.

To help catch usages of the old ANY, it actually will *only* take a
GROUP! for the time being...it does this by quoting.  When all the
old ANYs are known to be gone, it can take a parser that synthesizes
its product through whatever means.