Rethink @XXX combinators to synthesize literal values
In practice, working with COMPOSE revealed that it is difficult when
the only way to create literal values in UPARSE invovles GROUP!s.
That is contentious with trying to build rules via COMPOSE.

This makes the @ rules--in particular @ itself--produce literal values
out of thin air without groups.

The old behaviors of @ rules for matching values in variables "as is"
instead of as rules are put aside for later review.