Some typeset making automation, ANY-META-TYPE? etc.
The process for making typesets and words for them is only slightly
automated.  Some of this is for "a reason", which is that the types
are laid out in an order that makes optimizations possible that do
not require checking bitsets to tell if something is an ANY_GET_KIND()
or other such tests.  That doesn't make it a good reason (the
optimization may be irrelevant, with only a few like Is_Bindable()
actually mattering).  But at least it's a reason.

Some of the rest of the lack of automation is just because the code
is old and has evolved organically.  This takes a small step in
automation to add the typeset symbols instead of naming them in
%words.r, and gives new tests for ANY-THE-TYPE! etc.