Make appending the @ types remove the @
Working with the rule that evaluative values can't be appended unless
quoted makes for some difficulty APPENDing or KEEPing WORD!s, and it
makes some code awkward.  Since the number of types accepted is
somewhat narrow, this tries out the concept of making the inert
symbolic types have special behavior and permit appending.

    >> append [a b c] @d
    == [a b c d]

This provides an alternative to approaches like `[a b c] ^ 'd`, and
works particularly well when UPARSE is modified to treat @xxx values
as inert...because it doesn't have the option of 'd as that would
mean matching a literal d in the input.