Simple FAIL combinator
This is a very modest initial crack at a FAIL combinator, which takes
a THE-BLOCK! as a reason for failing that it converts to a string,
and reports the current parse input location as the problem source.

Describing the parse stack meaningfully in the WHERE is more complex.
This is only being added to try and pare down the size of the
Whitespace interpreter sample code posted on the forum, so tests and
further features will be added at a later date.