Make %prot-tls.r use dots for field access
While it's by no means secure, the usermode TLS protocol has some
interesting dialecting and represents a kind of sketch of methods
for a very "spec driven" implementation.

So continuing to keep it afloat and modernizing it is worthwhile.
This changes it to embrace the idea that slashes are used for
refinements, and dots are used for field selection.