Change rule for default REDUCE, add DENULL and REDUCE*
There has been contention over what the default behavior for REDUCE
should be with respect to NULL.  It can't be put in the block, so
the choices are to vaporize it...raise an error...or put "something
else" in the slot.

The Ren-C bias was to error.  However, due to users saying they liked
the idea of vaporization, that was tried for a while.

But the prescriptiveness for usage of NULL as a "step to more ornery
than opting out" in variables made the precariousness of this more
of a problem than before.  Especially with the rise of GET-BLOCK!
as a synonym for reduce, things like:

    min: null, max: _

    :[min max]  => [_]

...undermined the idea of using NULL to say a variable was "not set,
but not in a way that was as unfriendly as a BAD-WORD! isotope".

So this shifts it so that NULL makes ~null~ in REDUCE situations.  It
provides a DENULL operator to vanish nulls if that is truly the intent.

It also provides REDUCE* as a specialization of REDUCE with the
predicate of DENULL, for those who want the vanishing behavior easiliy
accessible.