New rules for COMPOSE splicing and NULL handling
This brings COMPOSE's ((splicing)) rules in sync with the rules of
APPEND; e.g. BLANK!s vanish, only BLOCK!s splice, anything evaluative
is an error, QUOTED!s drop a quote level and append as-is, other
ANY-INERT!s append as-is and all evaluative items give an error.

One small twist is that since the ((splicing)) indicator has shown the
user is prepared to accept that there's not a "one and only one"
element being put in the slot, that NULL is allowed to vaporize there.
Also, void isotopes vanish instead of erroring.

For the (single) form of splice, the rules now are such that NULL is
turned into ~null~, thus preserving the single slot invariant but
without raising an error.

To make this in sync with how quoted slots work in APPEND, appending
a quoted NULL will append ~null~:

    >> append [a b c] the '
    == [a b c ~null~]

Pure invisibles will vanish in either slot, but void isotopes only
vanish in the spliced case for now...since they have a more pronounced
tolerance for a variant number of items.  Consider why DO returns
void isotopes:

    >> x: do frame 10
    == ~void~  ; isotope
       ^-- we *never* expect 10 when calling DO like this

The same motivation that X does not want to be 10 can be tied to
why the (single) splice here should not vaporize:

    >> compose [x: (do frame)]
    ** Error: cannot splice void isotope in (...), use ((...))