Reimagine JOIN, eliminate JOIN-ALL and REJOIN
This changes JOIN to be non-evaluative, and to do a kind of "path and
tuple calculus" to ensure the coherence of paths and tuples that
are being joined.

JOIN-ALL is eliminated, instead making it possible to specify the
type of the JOIN result as the first argument:

    >> join get-path! [_ _ a / b _ _]
    == :a/b

REJOIN--which has long been marked for deletion--may someday be claimed
for an shorthand for JOIN/REDUCE, which would be able to JOIN and
REDUCE at the same time, lowering the intermediate series overhead.