New Positional and Named Argument Hybrid APPLY
Early on in Ren-C, the concept of moving away from positional
refinements led to the idea of "frames" being built in userspace.  The
first take on reinventing APPLY worked with names for all the
arguments, and ran a block of code bound into a frame made for the
function.

That idea was nice, and is today called APPLIQUE.  But not all apply
scenarios surrounded the idea of wanting to name all the arguments.
Some of them just wanted to have a clearer organization of the
parameters, putting the names of refinements next to the value for
those refinements instead of being so distant.  So it was more of an
organizational tool.

Having settled on the idea that /FOO will not have behavior in the
evaluator, and that it should be used for argument labels instead of
FOO:, a positional-labeled APPLY was ready to be written:

    >> apply :append [[a b c] [d e] /dup 2])
    == [a b c d e d e]

Any order for refinements is legal, and interstitial commas are legal:

    >> apply :append [/dup 2, [a b c] [d e]])
    == [a b c d e d e]

Because of the "refinements are their own arguments" philosophy (no
refinements take more than 1 arg), the process could be done where any
argument can be named via refinement.  Once named, an argument is
removed from consideration for positional fulfillment:

    >> apply :append [/value [d e] [a b c]]
    == [a b c d e]