TEMPORARY (?)  Allow writing of all attached variables.
The long term hope with the Sea of Words change is to have something
more like JavaScript's "strict mode", where writing to a variable
needs some sort of forward declaration or implicit "top level
declaration gathering" to permit it to write.  Otherwise, you can
easily make typos that create unwanted variables.

That's a very new model for Redbols, and so converting code to the
new rules will be a long road.  Getting Sea of Words committed so that
the experiment can be refined is a more realistic goal, so go ahead
and make it write to any variable regardless of its declaration status.

It's no worse than things used to be, and allows the other notable
aspects of Sea of Words to be exercised and examined.